Add username features
I don't want anyone to connect without a username, as it would mean that
one could silently join the chat and listen without active participants
knowing. To fix this I force the listener to send a username together with
the UUID. If the username is null, then the server generates a username.

I also added a bunch of functions for allowing a user to change its user-
name.